// src/app/common/model/code.model.ts
export interface Code {
    id: number;
    codeCd: string;
    codeD1: string;
    codeD2: string;
    codeD3: string;
    codeN1: string;
    codeN2: string;
    codeN3: string;
    codeS1: string;
    codeS2: string;
    codeS3: string;
    kindCode: string;
    remark: string;
    sort: number;
    used: boolean;
    codeName: string;
    codeNameEng: string;
}